requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
openai-whisper==20230314
torch==2.3.1
torchaudio==2.3.1
python-multipart==0.0.6
librosa==0.11.0